MacOS X compilation support.
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libhxcfe/trunk@429 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ libhxcfe/sources/loaders/cpcdsk_loader/cpcdsk_writer.c
@@ -107,7 +107,7 @@ int CPCDSK_libWrite_DiskFile(HXCFLOPPYEMULATOR* floppycontext,FLOPPY * floppy,ch
 		{
 			for(i=0;i<(int)floppy->floppyNumberOfSide;i++)
 			{
-				sprintf(tmp_str,"track:%.2d:%d file offset:0x%.6x, sectors: ",j,i,ftell(cpcdskfile));
+				sprintf(tmp_str,"track:%.2d:%d file offset:0x%.6x, sectors: ",j,i,(unsigned int)ftell(cpcdskfile));
 
 				log_str=0;
 				log_str=realloc(log_str,strlen(tmp_str)+1);
